split DocumentEditor/insert-menu code into shared/ui
